First commit in order to make IScala more Spark friendly
